CHAT-67 Implement an active button state - final
